requirements.txt
psycopg2==2.5.1
django-bootstrap-toolkit==2.15.0
httpagentparser==1.2.2
